placessidebar: add Other Locations item
Places sidebar is a widget that enabled the user to
select XDG directories, bookmarks and mounted network
locations, as well as manages permanent and removable
devices.

As part of the ongoing effort to modernize GNOME plataform,
places sidebar is supposed to handle only removable
devices, as well as mounted networks, bookmarks and XDG
directories, and delegate the management of permanent
devices such as hard drive partitions to other widgets.

To delegate it, add an "Other Locations..." item to
notify when the permanent devices manager is required.
Besides that, don't show these fixes devices on the
sidebar itself, as they are supposed to be delegated.